requirements.txt
streamlit 
firebase_admin 
google-cloud-firestore 
streamlit-on-Hover-tabs 
streamlit-chat 
langchain 
